# feat: Implement high-performance client-side audio extraction from video files (#22)