--- a/backend/Dockerfile.prod
+++ b/backend/Dockerfile.prod
@@ -1,4 +1,4 @@
-FROM python:3.10.12-slim-bullseye
+FROM python:3.12.11-slim-bullseye
 
 WORKDIR /app