--- a/backend/app/models/prompt.py
+++ b/backend/app/models/prompt.py
@@ -1,6 +1,7 @@
 """
 SQLAlchemy models for AI summarization prompt management
 """
+
 import uuid as uuid_pkg
 
 from sqlalchemy import Boolean
@@ -28,7 +29,9 @@ class SummaryPrompt(Base):
     __tablename__ = "summary_prompt"
 
     id = Column(Integer, primary_key=True, index=True)
-    uuid = Column(UUID(as_uuid=True), unique=True, nullable=False, default=uuid_pkg.uuid4, index=True)
+    uuid = Column(
+        UUID(as_uuid=True), unique=True, nullable=False, default=uuid_pkg.uuid4, index=True
+    )
     name = Column(String(255), nullable=False, index=True)
     description = Column(Text, nullable=True)
     prompt_text = Column(Text, nullable=False)
@@ -55,7 +58,9 @@ class UserSetting(Base):
     __tablename__ = "user_setting"
 
     id = Column(Integer, primary_key=True, index=True)
-    uuid = Column(UUID(as_uuid=True), unique=True, nullable=False, default=uuid_pkg.uuid4, index=True)
+    uuid = Column(
+        UUID(as_uuid=True), unique=True, nullable=False, default=uuid_pkg.uuid4, index=True
+    )
     user_id = Column(Integer, ForeignKey("user.id", ondelete="CASCADE"), nullable=False, index=True)
     setting_key = Column(String(100), nullable=False, index=True)
     setting_value = Column(Text, nullable=True)  # Can store JSON or simple values