… backend/app/api/endpoints/user_settings.py
… backend/app/models/media.py
… backend/app/models/user.py
… backend/app/models/user_llm_settings.py
… backend/app/schemas/base.py
… backend/app/schemas/media.py
… backend/app/schemas/user.py
… backend/app/services/llm_service.py
… backend/app/services/opensearch_service.py
… backend/app/services/opensearch_summary_service.py
… backend/app/services/profile_embedding_service.py
… backend/app/services/smart_speaker_suggestion_service.py
… backend/app/services/speaker_matching_service.py
… backend/app/services/topic_extraction_service.py
… backend/app/tasks/topic_extraction.py
… backend/app/tasks/transcription/core.py
… backend/app/tasks/youtube_processing.py
… frontend/package.json
… frontend/src/components/AudioExtractionModal.svelte
… frontend/src/components/BulkAudioExtractionModal.svelte
… frontend/src/components/FileUploader.svelte
… frontend/src/components/NotificationsPanel.svelte
… frontend/src/components/VideoPlayer.svelte
… frontend/src/components/settings/AudioExtractionSettings.svelte
… frontend/src/lib/api/audioExtractionSettings.ts
… frontend/src/lib/services/audioExtractionService.ts
… frontend/src/lib/services/uploadService.ts
… frontend/src/lib/types/audioExtraction.ts
… frontend/src/lib/utils/metadataMapper.ts
… frontend/src/routes/UserSettings.svelte